Create Classname.hbm.xml
Hibernate maps Pojo class name to Table. If Class name and Table name are same, writing XML in table name is optional and if Variable name and Table Column name are same mapping file Column name is Optional. All XML files are extended to . Xml but, in hibernate mapping, the file name is extended to Xxx .hbm.xml. Following is the syntax of writing a mapping file in Hibernate.